Move file related funcs to RecordingsManager
